# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RSR Media API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  passcode: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  draft: "draft",
  published: "published",
} as const;

export interface Report {
  id: string;
  reportNumber: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  slug: string;
  category: string;
  date: string;
  description: string;
  fullDescription: string;
  tags: string[];
  /** @nullable */
  sourceDocument?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  pdfStorageKey?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroImageStorageKey?: string | null;
  /** @nullable */
  shopifyUrl?: string | null;
  status: ReportStatus;
  featured: boolean;
  createdAt: string;
  updatedAt: string;
}

export type ReportInputStatus =
  (typeof ReportInputStatus)[keyof typeof ReportInputStatus];

export const ReportInputStatus = {
  draft: "draft",
  published: "published",
} as const;

export interface ReportInput {
  /** @minLength 1 */
  reportNumber: string;
  /** @minLength 1 */
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  category: string;
  /** @nullable */
  date?: string | null;
  description?: string;
  fullDescription?: string;
  tags?: string[];
  /** @nullable */
  sourceDocument?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  pdfStorageKey?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroImageStorageKey?: string | null;
  /** @nullable */
  shopifyUrl?: string | null;
  status: ReportInputStatus;
  featured?: boolean;
}

export type ReportUpdateStatus =
  (typeof ReportUpdateStatus)[keyof typeof ReportUpdateStatus];

export const ReportUpdateStatus = {
  draft: "draft",
  published: "published",
} as const;

export interface ReportUpdate {
  /** @minLength 1 */
  reportNumber?: string;
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  subtitle?: string | null;
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  category?: string;
  /** @nullable */
  date?: string | null;
  description?: string;
  fullDescription?: string;
  tags?: string[];
  /** @nullable */
  sourceDocument?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  pdfStorageKey?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroImageStorageKey?: string | null;
  /** @nullable */
  shopifyUrl?: string | null;
  status?: ReportUpdateStatus;
  featured?: boolean;
}

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListReportsParams = {
  includeDrafts?: boolean;
};