# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RSR Media API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: admin
  - name: reports
  - name: storage
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthStatus" }

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Verify admin passcode and start session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminLoginInput" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSession" }
        "401":
          description: Invalid passcode
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: End admin session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSession" }

  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Get current admin session state
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSession" }

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports (published only for public, all for admin)
      parameters:
        - name: includeDrafts
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a report (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Slug already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /reports/{slug}:
    get:
      operationId: getReportBySlug
      tags: [reports]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /reports/id/{id}:
    patch:
      operationId: updateReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Slug conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      operationId: deleteReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadRequest" }
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadUrlResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }

    AdminLoginInput:
      type: object
      required: [passcode]
      properties:
        passcode: { type: string, minLength: 1 }

    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }

    Report:
      type: object
      required:
        - id
        - reportNumber
        - title
        - slug
        - category
        - status
        - featured
        - date
        - description
        - fullDescription
        - tags
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        reportNumber: { type: string }
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        slug: { type: string }
        category: { type: string }
        date: { type: string, format: date-time }
        description: { type: string }
        fullDescription: { type: string }
        tags:
          type: array
          items: { type: string }
        sourceDocument: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        pdfUrl: { type: ["string", "null"] }
        pdfStorageKey: { type: ["string", "null"] }
        heroImageUrl: { type: ["string", "null"] }
        heroImageStorageKey: { type: ["string", "null"] }
        shopifyUrl: { type: ["string", "null"] }
        status: { type: string, enum: [draft, published] }
        featured: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    ReportInput:
      type: object
      required: [reportNumber, title, slug, category, status]
      properties:
        reportNumber: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        subtitle: { type: ["string", "null"] }
        slug: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        date: { type: ["string", "null"], format: date-time }
        description: { type: string }
        fullDescription: { type: string }
        tags:
          type: array
          items: { type: string }
        sourceDocument: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        pdfUrl: { type: ["string", "null"] }
        pdfStorageKey: { type: ["string", "null"] }
        heroImageUrl: { type: ["string", "null"] }
        heroImageStorageKey: { type: ["string", "null"] }
        shopifyUrl: { type: ["string", "null"] }
        status: { type: string, enum: [draft, published] }
        featured: { type: boolean }

    ReportUpdate:
      type: object
      properties:
        reportNumber: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        subtitle: { type: ["string", "null"] }
        slug: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        date: { type: ["string", "null"], format: date-time }
        description: { type: string }
        fullDescription: { type: string }
        tags:
          type: array
          items: { type: string }
        sourceDocument: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        pdfUrl: { type: ["string", "null"] }
        pdfStorageKey: { type: ["string", "null"] }
        heroImageUrl: { type: ["string", "null"] }
        heroImageStorageKey: { type: ["string", "null"] }
        shopifyUrl: { type: ["string", "null"] }
        status: { type: string, enum: [draft, published] }
        featured: { type: boolean }

    UploadRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 0 }
        contentType: { type: string, minLength: 1 }

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }