# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RSR Media API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatus } from "./reportStatus";

export interface Report {
  id: string;
  reportNumber: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  slug: string;
  category: string;
  date: Date;
  description: string;
  fullDescription: string;
  tags: string[];
  /** @nullable */
  sourceDocument?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  pdfStorageKey?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroImageStorageKey?: string | null;
  /** @nullable */
  shopifyUrl?: string | null;
  status: ReportStatus;
  featured: boolean;
  createdAt: Date;
  updatedAt: Date;
}