# lib/api-zod/src/generated/types/reportUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RSR Media API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportUpdateStatus } from "./reportUpdateStatus";

export interface ReportUpdate {
  /** @minLength 1 */
  reportNumber?: string;
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  subtitle?: string | null;
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  category?: string;
  /** @nullable */
  date?: Date | null;
  description?: string;
  fullDescription?: string;
  tags?: string[];
  /** @nullable */
  sourceDocument?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  pdfStorageKey?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroImageStorageKey?: string | null;
  /** @nullable */
  shopifyUrl?: string | null;
  status?: ReportUpdateStatus;
  featured?: boolean;
}