# lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RSR Media API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}